Revert "Abstracted the common feed functionality"
This reverts PR #750
